General maintenance of int/size_t vars, and new logic to search for keys.
- use less size_t, use more int. Most of the size_t-s were unnecessary.
- use const variables for function parameters, where possible.
- use unsigned variables for function parameters and elsewhere, where it is justified.
- remove unnecessary (long) castings. I think lint messed with these a long time ago...
- use xmlStrlen() with xmlChar-s, and don't use needless type casting.
- fine-tune checks for spice, explicitly check if it is greater than 0.
- find_key() logic rewrite. eg.: don't start the index var. from -1 ... .
- when dealing with value lines, precalculate their lengths, and don't call (xml)strlen every time in a loop.

+++ cmd_quit.c
@@ -83,7 +83,7 @@ cmd_quit(const char *e_line, command *commands)
 			line = strdup(e_line);
 
 #ifndef _READLINE
-			line[(long)(strlen(line) - 1)] = '\0';	/* remove the newline */
+			line[strlen(line) - 1] = '\0';	/* remove the newline */
 #endif
 
 			if (strcasecmp(line, "yes") == 0  ||  strcasecmp(line, "y") == 0) {
